`Coalesce` as helper
