+++ src/interface/config.rs
@@ -14,7 +14,7 @@ use crate::{
     interface::template::Template,
 };
 
-use super::helper::http_serde_priv;
+use super::helper::{coalesce::Coalesce, http_serde_priv};
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
 #[serde(deny_unknown_fields, rename_all = "kebab-case")]
@@ -160,29 +160,6 @@ impl<Q: Coalesce, P: Coalesce> Coalesce for Setting<Q, P> {
             timeout: self.timeout.or(other.timeout),
             response: self.response.coalesce(&other.response),
         }
-    }
-}
-
-pub trait Coalesce<O = Self> {
-    fn coalesce(self, other: &O) -> Self;
-}
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct Coalesced<T, O> {
-    base: T,
-    coalesced: Vec<O>,
-}
-impl<T: Clone + Coalesce<O>, O> Coalesced<T, O> {
-    pub fn new(base: T, coalesced: Vec<O>) -> Self {
-        Self { base, coalesced }
-    }
-    pub fn tuple(base: T, other: O) -> Self {
-        Self::new(base, vec![other])
-    }
-    pub fn coalesce(&self) -> T {
-        self.coalesced.iter().fold(self.base.clone(), |acc, x| acc.coalesce(x))
-    }
-    pub fn base(&self) -> &T {
-        &self.base
     }
 }
